meson/nanosvg: add support for linking against system's nanosvg
Nanosvg upstream is header only, and does not support/provide a
library, neither shared nor static.

However, most distributions still provide a nanosvg package with
either a static library, or shared. Arch does the latter.

So, let's support building against a system provided nanosvg library.

The default is still to use bundled version, but this can be changed
with

    -Dsystem-nanosvg=enabled

when configuring the build tree.